Extract resumeBuild out of submissions.ts
`resumeBuild` starts the next round on a job that already exists, keeping the
thread and the history it has. Five modules already take it as a dep — the
reconciler, feedback, handoff+seal, and two callers still inside submissions.ts
— so it was the most-depended-on function in the file. It moves to
`creation/resume-build.ts` behind `createResumeBuild`, along with the three
things that exist only to describe its outcome: `ResumeFailureReason`,
`ResumeOutcome` and `classifyResumeFailure`.

Nothing outside submissions.ts imported any of those three, despite all being
exported. That is what made the extraction clean, and what makes the follow-on
possible:

`feedback-routes.ts` and `handoff-seal-routes.ts` each carried a structural
stand-in for `ResumeOutcome`, with a comment explaining that the real type
lived in submissions.ts and importing it would be a cycle. `resume-build.ts` is
a leaf, so both now import the real type and the workaround is gone. The two
comments would otherwise have been left asserting something no longer true.

`resumeBuild` does not call `dispatchBuild`, so this splits off cleanly on its
own; `dispatchBuild` and `redispatchQueuedJob` follow separately.

Verified by diffing the body against master with comments and blank lines
stripped: byte-identical, 121 code lines. `classifyResumeFailure` is identical
too.

submissions.ts: 2092 -> 1917 lines, ceiling resealed.
